Consider interface of altiria-mock the SMS iface
The functionality Kyuubi requires from Altiria and any other SMS service
provider is similar: just send SMS messages. In this sense, the API
implemented in `altiria-mock` module may be considered not a wrapper
of Altiria but a wrapper of any SMS service provider.

This change assumes that fact. The modules are renamed from `altiria-*` to
`sms-*`.